Refactor History.py -> history.py
