+++ medusa/server/api/v1/core.py
@@ -52,7 +52,7 @@ from ....media.ShowNetworkLogo import ShowNetworkLogo
 from ....media.ShowPoster import ShowPoster
 from ....search.queue import BacklogQueueItem, ForcedSearchQueueItem
 from ....show.ComingEpisodes import ComingEpisodes
-from ....show.History import History
+from ....show.history import History
 from ....show.show import Show
 from ....system.restart import Restart
 from ....system.shutdown import Shutdown
