+++ pytest.ini
@@ -209,7 +209,7 @@ flake8-ignore =
     medusa/recompiled/__init__.py D104
     medusa/recompiled/tags.py D100
     medusa/show/ComingEpisodes.py D100 D102 D202 D204 D205 D400 E501
-    medusa/show/History.py D100 D101 D102 D200 D202 D204 D205 D400 D401
+    medusa/show/history.py D100 D101 D102 D200 D202 D204 D205 D400 D401
     medusa/show/__init__.py D104
     medusa/show/recommendations/anidb.py D100 D101 D102 D202 D400 E128 E501
     medusa/show/recommendations/imdb.py D100 D101 D102 D202 D205 D400 D401 E501
