+++ views/history.mako
@@ -9,7 +9,7 @@
     from medusa.sbdatetime import sbdatetime
     from medusa.common import SKIPPED, WANTED, UNAIRED, ARCHIVED, IGNORED, SNATCHED, SNATCHED_PROPER, SNATCHED_BEST, FAILED, DOWNLOADED, SUBTITLED
     from medusa.common import Quality, statusStrings, Overview
-    from medusa.show.History import History
+    from medusa.show.history import History
     from medusa.providers.GenericProvider import GenericProvider
 %>
 <%block name="content">
